JeeObjectEntity
Tester ce qui est lié aux objets pour voir si tout va bien
