--- src/Model/Entity/Object.php
@@ -1,178 +1,0 @@
-<?php
-/* This file is part of NextDom Software.
- *
- * NextDom is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * NextDom Software is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with NextDom Software. If not, see <http://www.gnu.org/licenses/>.
- */
-
-namespace NextDom\Model\Entity;
-
-/**
- * Object
- *
- * @ORM\Table(name="object", uniqueConstraints={@ORM\UniqueConstraint(name="name_UNIQUE", columns={"name"})}, indexes={@ORM\Index(name="fk_object_object1_idx1", columns={"father_id"}), @ORM\Index(name="position", columns={"position"})})
- * @ORM\Entity
- */
-class Object
-{
-
-    /**
-     * @var string
-     *
-     * @ORM\Column(name="name", type="string", length=45, nullable=false)
-     */
-    private $name;
-
-    /**
-     * @var boolean
-     *
-     * @ORM\Column(name="isVisible", type="boolean", nullable=true)
-     */
-    private $isvisible;
-
-    /**
-     * @var integer
-     *
-     * @ORM\Column(name="position", type="integer", nullable=true)
-     */
-    private $position;
-
-    /**
-     * @var string
-     *
-     * @ORM\Column(name="configuration", type="text", length=65535, nullable=true)
-     */
-    private $configuration;
-
-    /**
-     * @var string
-     *
-     * @ORM\Column(name="display", type="text", length=65535, nullable=true)
-     */
-    private $display;
-
-    /**
-     * @var string
-     *
-     * @ORM\Column(name="image", type="text", length=16777215, nullable=true)
-     */
-    private $image;
-
-    /**
-     * @var integer
-     *
-     * @ORM\Column(name="id", type="integer")
-     * @ORM\Id
-     * @ORM\GeneratedValue(strategy="IDENTITY")
-     */
-    private $id;
-
-    /**
-     * @var \NextDom\Model\Entity\Object
-     *
-     * @ORM\ManyToOne(targetEntity="NextDom\Model\Entity\Object")
-     * @ORM\JoinColumns({
-     *   @ORM\JoinColumn(name="father_id", referencedColumnName="id")
-     * })
-     */
-    private $father;
-
-    public function getName()
-    {
-        return $this->name;
-    }
-
-    public function getIsvisible()
-    {
-        return $this->isvisible;
-    }
-
-    public function getPosition()
-    {
-        return $this->position;
-    }
-
-    public function getConfiguration()
-    {
-        return $this->configuration;
-    }
-
-    public function getDisplay()
-    {
-        return $this->display;
-    }
-
-    public function getImage()
-    {
-        return $this->image;
-    }
-
-    public function getId()
-    {
-        return $this->id;
-    }
-
-    public function getFather(): \NextDom\Model\Entity\Object
-    {
-        return $this->father;
-    }
-
-    public function setName($name)
-    {
-        $this->name = $name;
-        return $this;
-    }
-
-    public function setIsvisible($isvisible)
-    {
-        $this->isvisible = $isvisible;
-        return $this;
-    }
-
-    public function setPosition($position)
-    {
-        $this->position = $position;
-        return $this;
-    }
-
-    public function setConfiguration($configuration)
-    {
-        $this->configuration = $configuration;
-        return $this;
-    }
-
-    public function setDisplay($display)
-    {
-        $this->display = $display;
-        return $this;
-    }
-
-    public function setImage($image)
-    {
-        $this->image = $image;
-        return $this;
-    }
-
-    public function setId($id)
-    {
-        $this->id = $id;
-        return $this;
-    }
-
-    public function setFather(\NextDom\Model\Entity\Object $father)
-    {
-        $this->father = $father;
-        return $this;
-    }
-
-}
